requirements.txt
aiohttp==3.10.5
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
python-dotenv==1.0.1
requests==2.32.3
simple-websocket==1.0.0
bcrypt==4.2.0